day22

+++ controllers/ProductController.js
@@ -120,38 +120,46 @@ app.post('/uploadFromExcel', (req, res) => {
     try {
         const fileExcel = req.files.fileExcel;
 
-        fileExcel.mv('./uploads/' + fileExcel.name, async (err) => {
-            if (err) throw err;
+        if (fileExcel != undefined) {
+            if (fileExcel != null) {
+                fileExcel.mv('./uploads/' + fileExcel.name, async (err) => {
+                    if (err) throw err;
 
-            // read from file and insert to database
-            const workbook = new exceljs.Workbook();
-            await workbook.xlsx.readFile('./uploads/' + fileExcel.name);
+                    // read from file and insert to database
+                    const workbook = new exceljs.Workbook();
+                    await workbook.xlsx.readFile('./uploads/' + fileExcel.name);
 
-            const ws = workbook.getWorksheet(1);
+                    const ws = workbook.getWorksheet(1);
 
-            for (let i = 2; i <= ws.rowCount; i++) {
-                const name = ws.getRow(i).getCell(1).value ?? ""; // if null of undefined return ""
-                const cost = ws.getRow(i).getCell(2).value ?? 0; // if null of undefined return 0
-                const price = ws.getRow(i).getCell(3).value ?? 0; // if null of undefined return 0
+                    for (let i = 2; i <= ws.rowCount; i++) {
+                        const name = ws.getRow(i).getCell(1).value ?? ""; // if null of undefined return ""
+                        const cost = ws.getRow(i).getCell(2).value ?? 0; // if null of undefined return 0
+                        const price = ws.getRow(i).getCell(3).value ?? 0; // if null of undefined return 0
 
-                if (name != "" && cost >= 0 && price >= 0) {
-                    await prisma.product.create({
-                        data: {
-                            name: name,
-                            cost: cost,
-                            price: price,
-                            img: ''
+                        if (name != "" && cost >= 0 && price >= 0) {
+                            await prisma.product.create({
+                                data: {
+                                    name: name,
+                                    cost: cost,
+                                    price: price,
+                                    img: ''
+                                }
+                            })
                         }
-                    })
-                }
+                    }
+
+                    // remove file from server
+                    const fs = require('fs');
+                    await fs.unlinkSync('./uploads/' + fileExcel.name);
+
+                    res.send({ message: 'success' })
+                })
+            } else {
+                res.status(500).send({ message: 'fileExcel is null' });
             }
-
-            // remove file from server
-            const fs = require('fs');
-            await fs.unlinkSync('./uploads/' + fileExcel.name);
-
-            res.send({ message: 'success' })
-        })
+        } else {
+            res.status(500).send({ message: 'fileExcel is undefined' });
+        }
     } catch (e) {
         res.status(500).send({ error: e.message });
     }
